Split the base class for checks
This allows for a base class that requires a typed check and a a sub
class that implements the typed test and requires the non typed test as
it was before.
