+++ checkov/arm/base_resource_check.py
@@ -1,18 +1,13 @@
 from abc import abstractmethod
 
-from checkov.common.checks.base_check import BaseCheck
-from checkov.arm.registry import arm_registry
+from .typed_base_resource_check import TypedBaseResourceCheck
 
-class BaseResourceCheck(BaseCheck):
-    def __init__(self, name, id, categories, supported_resources):
-        super().__init__(name=name, id=id, categories=categories, supported_entities=supported_resources,
-                         block_type="resource")
-        self.supported_resources = supported_resources
-        arm_registry.register(self)
+
+class BaseResourceCheck(TypedBaseResourceCheck):
+
+    def typed_scan_resource_conf(self, conf, entity_type):
+        return self.scan_resource_conf(conf)
 
     @abstractmethod
     def scan_resource_conf(self, conf):
         raise NotImplementedError()
-
-    def scan_entity_conf(self, conf, entity_type):
-        return self.scan_resource_conf(conf)
