+++ checkov/kubernetes/base_spec_check.py
@@ -1,23 +1,12 @@
 from abc import abstractmethod
 
-from checkov.common.checks.base_check import BaseCheck
-from checkov.kubernetes.registry import registry
+from .typed_base_spec_check import TypedBaseK8Check
 
 
-class BaseK8Check(BaseCheck):
-    def __init__(self, name, id, categories, supported_entities):
-        super().__init__(name=name, id=id, categories=categories, supported_entities=supported_entities,
-                         block_type="k8")
-        self.supported_specs = supported_entities
-        registry.register(self)
+class BaseK8Check(TypedBaseK8Check):
+    def typed_scan_spec_conf(self, conf, entity_type):
+        return self.scan_spec_conf(conf)
 
     @abstractmethod
     def scan_spec_conf(self, conf):
         raise NotImplementedError()
-
-    def scan_entity_conf(self, conf, entity_type):
-        return self.scan_spec_conf(conf)
-
-    @abstractmethod
-    def get_resource_id(self, conf):
-        pass
