+++ checkov/terraform/checks/data/base_check.py
@@ -1,19 +1,13 @@
 from abc import abstractmethod
 
-from checkov.terraform.checks.data.registry import data_registry
-from checkov.common.checks.base_check import BaseCheck
+from .typed_base_data_check import TypedBaseDataCheck
 
 
-class BaseDataCheck(BaseCheck):
-    def __init__(self, name, id, categories, supported_data):
-        super().__init__(name=name, id=id, categories=categories, supported_entities=supported_data,
-                         block_type="data")
-        self.supported_data = supported_data
-        data_registry.register(self)
+class BaseDataCheck(TypedBaseDataCheck):
+
+    def typed_scan_data_conf(self, conf, entity_type):
+        return self.scan_data_conf(conf)
 
     @abstractmethod
     def scan_data_conf(self, conf):
         raise NotImplementedError()
-
-    def scan_entity_conf(self, conf, entity_type):
-        return self.scan_data_conf(conf)
