+++ checkov/terraform/checks/module/base_module_check.py
@@ -1,19 +1,14 @@
 from abc import abstractmethod
 
-from checkov.common.checks.base_check import BaseCheck
-from checkov.terraform.checks.module.registry import module_registry
+from .typed_base_module_check import TypedBaseModuleCheck
 
 
-class BaseModuleCheck(BaseCheck):
-    def __init__(self, name, id, categories, supported_resources):
-        super().__init__(name=name, id=id, categories=categories, supported_entities=supported_resources,
-                         block_type="module")
-        self.supported_resources = supported_resources
-        module_registry.register(self)
+class BaseModuleCheck(TypedBaseModuleCheck):
+
+    def typed_scan_module_conf(self, conf, entity_type):
+        return self.scan_resource_conf(conf)
 
     @abstractmethod
     def scan_resource_conf(self, conf):
+        # TODO rename to scan_module_conf
         raise NotImplementedError()
-
-    def scan_entity_conf(self, conf, entity_type):
-        return self.scan_resource_conf(conf)
