+++ checkov/terraform/checks/provider/base_check.py
@@ -1,19 +1,12 @@
 from abc import abstractmethod
 
-from checkov.terraform.checks.provider.registry import provider_registry
-from checkov.common.checks.base_check import BaseCheck
+from .typed_base_provider_check import TypedBaseProviderCheck
 
 
-class BaseProviderCheck(BaseCheck):
-    def __init__(self, name, id, categories, supported_provider):
-        super().__init__(name=name, id=id, categories=categories, supported_entities=supported_provider,
-                         block_type="provider")
-        self.supported_provider = supported_provider
-        provider_registry.register(self)
+class BaseProviderCheck(TypedBaseProviderCheck):
+    def typed_scan_provider_conf(self, conf, entity_type):
+        return self.scan_provider_conf(conf)
 
     @abstractmethod
     def scan_provider_conf(self, conf):
         raise NotImplementedError()
-
-    def scan_entity_conf(self, conf, entity_type):
-        return self.scan_provider_conf(conf)
